requirements.txt
# 
langchain-ollama
langchain
ollama


#
googlesearch-python
google




